Corrigindo padrão fabrica

+++ back/cafeteria/src/main/java/com/spindola/cafeteria/domain/model/ItemPedidoModel.java
@@ -1,6 +1,6 @@
 package com.spindola.cafeteria.domain.model;
 
-import java.util.List;
+import java.math.BigDecimal;
 
 import com.spindola.cafeteria.domain.model.interfaces.IProduto;
 
@@ -8,7 +8,7 @@ public class ItemPedidoModel {
     private Long id;
     //private Pedido pedido;
     private IProduto cafeBase;
-    private double valorTotalItem;
+    private BigDecimal valorTotalItem;
 
     public ItemPedidoModel() {
     }
@@ -21,7 +21,7 @@ public class ItemPedidoModel {
         return cafeBase;
     }
 
-    public double getValorTotalItem() {
+    public BigDecimal getValorTotalItem() {
         return valorTotalItem;
     }
 
@@ -33,7 +33,7 @@ public class ItemPedidoModel {
         this.cafeBase = cafeBase;
     }
 
-    public void setValorTotalItem(double valorTotalItem) {
+    public void setValorTotalItem(BigDecimal valorTotalItem) {
         this.valorTotalItem = valorTotalItem;
     }    
 }
